requirements.txt
Pillow==10.4.0
pystray==0.19.5
pywin32==306
